requirements.txt
langchain
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit
streamlit-authenticator
faiss-cpu==1.7.4
altair<5
tiktoken
unstructured
tabulate
pdf2image
watchdog
openai
psutil
